First working version of terminal connector
