+++ doc/Plugin-Author's-Guide.md
@@ -21,6 +21,7 @@ Table of Contents
     * [Authorization Plugins](#authorization-plugins)
     * [Elevation Plugins](#elevation-plugins)
     * [Other Reserved Commands](#other-reserved-commands)
+  * [Using the Terminal Connector](#using-the-terminal-connector)
   * [Getting Started](#getting-started)
     * [Starting from a Sample Plugin](#starting-from-a-sample-plugin)
     * [Using Boilerplate Code](#using-boilerplate-code)
@@ -105,6 +106,31 @@ In addition to the `configure` command, which instructs a plugin to dump it's de
 * `init` - During startup and reload, the robot will call external plugins with a command argument of `init`. Since all environment variables for the robot are set at that point, it would be possible to e.g. save a robot data structure that could be loaded and used in a cron job.
 * `event` - This command is reserved for future use with e.g. user presence change & channel join/leave events
 * `catchall` - Plugins with `CatchAll: true` will be called for commands directed at the robot that don't match a command plugin. Normally these are handled by the compiled-in `help` plugin, but administrators could override that setting and provide their own plugin with `CatchAll: true`. Note that having multiple such plugins is probably a bad idea.
+
+# Using the Terminal Connector
+Interacting with your bot in a chat app might not always be convenient or fast; to simplify
+testing and plugin development, **Gopherbot** includes a terminal connector, with a sample
+configuration in the `termcfg` directory. You'll want to copy the directory and modify
+it for your own use (mainly configuring the plugins you're developing), but it can be used
+by using the `-c <configdir>` option:
+```
+[./gopherbot]$ go build
+[./gopherbot]$ ./gopherbot -l /tmp/bot.log -c termcfg/
+Terminal connector running; Use '|C<channel>' to change channel, or '|U<user>' to change user
+c:general/u:alice -> floyd, info
+c:general/u:alice -> 
+general: Here's some information about my running environment:
+The hostname for the server I'm running on is: bot.example.org
+My install directory is: /home/username/go/src/github.com/lnxjedi/gopherbot
+My local configuration directory is: termcfg/
+My software version is: Gopherbot v1.0.1-snapshot, commit: (manual build)
+My alias is: ;
+The administrators for this robot are: alice
+c:general/u:alice -> ;quit
+c:general/u:alice -> 
+general: @alice Sayonara!
+[./gopherbot]$
+```
 
 # Getting Started
 ## Starting from a Sample Plugin
