+++ termcfg/brain/README.txt
@@ -1,0 +1,3 @@
+Until other plugins are developed, Gopherbot only supports a file-based
+brain for storing it's memories. Be sure <local config dir>/brain exists
+and is writable by the user the robot runs as.
