+++ termcfg/conf/plugins/duo.yaml.sample
@@ -1,0 +1,13 @@
+---
+## Configuration for Duo two-factor authentication. If your organization uses
+## Duo, you can obtain an IKey, SKey and Host for use with the auth api.
+Disabled: true
+# Config:
+## How long elevation lasts
+#   TimeoutSeconds: 7200
+## When 'idle', the timer resets on every elevated command
+#   TimeoutType: idle # or absolute
+#   DuoIKey: <YourIKey>
+#   DuoSKey: <YourSKey>
+#   DuoHost: <YourDuoHost>
+#   DuoUserString: emailUser
